seperated to different files and made a class out of bot
